+++ firestore-counter/functions/lib/controller.js
@@ -191,7 +191,7 @@ class ShardedCounterController {
                     yield t.get(this.controllerDocRef);
                 }
                 catch (err) {
-                    console.log("Failed to read controler doc " + this.controllerDocRef.path);
+                    console.log("Failed to read controller doc " + this.controllerDocRef.path);
                     throw Error("Failed to read controller doc.");
                 }
                 // Read all workers' metadata and construct sharding info based on collected stats.
@@ -201,7 +201,7 @@ class ShardedCounterController {
                 }
                 catch (err) {
                     console.log("Failed to read worker docs.", err);
-                    throw Error("Failed to reqad worker docs.");
+                    throw Error("Failed to read worker docs.");
                 }
                 let shardingInfo = yield Promise.all(query.docs.map((worker) => __awaiter(this, void 0, void 0, function* () {
                     const slice = worker.get("slice");