+++ firestore-counter/functions/src/worker.ts
@@ -29,7 +29,7 @@ import { Aggregator, NumericUpdate } from "./aggregator";
 import { FieldValue } from "@google-cloud/firestore";
 import * as uuid from "uuid";
 
-const SHARDS_LIMIT = 499;
+const SHARDS_LIMIT = 100;
 const WORKER_TIMEOUT_MS = 45000;
 
 interface WorkerMetadata {
@@ -50,7 +50,7 @@ interface WorkerMetadata {
  *
  * Workers avoid double scheduling and overruns by including their metadata documents in every
  * aggregation transaction. If metadata changes underneath, transaction fails, worker detects that
- * and terminates immmediately.
+ * and terminates immediately.
  */
 export class ShardedCounterWorker {
   private db: firestore.Firestore;